Add basic wordle grid, add user interactivity
This commit creates the foundation for what will become the
wordle game. Error checking has been implimented to see if the user
left any of the squares blank. The boxes will automatically fill
as the user types their guess. To submit a guess the user can press
the enter key. The guess grading system is not yet implimented.