Add game logic and change square background colors according to grade
This commit adds the game logic that grades the users guess.
If the letter is correct but in the wrong place it will be highlighted
with a yellow background. If the letter is in the right place and
correct it will be highlighted green. If the letter is not a match
at all it will be highlighted grey. The game will determine after
each guess if the user won by checking to see if all the letters
are highlighted green. In addition to this some code has been removed
to make the game more lightweight.